refactor: extract checkpoint proposal handling to ProposalHandler
Renames BlockProposalHandler to ProposalHandler and adds checkpoint
proposal handling, so non-validator nodes with blob upload capability
can validate checkpoint proposals and upload blobs to the filestore.

The validator client now delegates checkpoint validation and blob
upload to the ProposalHandler instead of doing it inline.
