+++ yarn-project/validator-client/src/index.ts
@@ -1,4 +1,4 @@
-export * from './block_proposal_handler.js';
+export * from './proposal_handler.js';
 export * from './checkpoint_builder.js';
 export * from './config.js';
 export * from './factory.js';
